+++ Gemfile
@@ -1,4 +1,4 @@
 source 'https://rubygems.org'
 
-# Specify your gem's dependencies in apartment-sidekiq.gemspec
+# Specify your gem's dependencies in innkeeper-sidekiq.gemspec
 gemspec
