+++ innkeeper-sidekiq.gemspec
@@ -1,16 +1,16 @@
 # coding: utf-8
 lib = File.expand_path('../lib', __FILE__)
 $LOAD_PATH.unshift(lib) unless $LOAD_PATH.include?(lib)
-require 'apartment/sidekiq/version'
+require 'innkeeper/sidekiq/version'
 
 Gem::Specification.new do |spec|
-  spec.name          = "apartment-sidekiq"
-  spec.version       = Apartment::Sidekiq::VERSION
-  spec.authors       = ["Brad Robertson"]
-  spec.email         = ["brad@influitive.com"]
+  spec.name          = "innkeeper-sidekiq"
+  spec.version       = Innkeeper::Sidekiq::VERSION
+  spec.authors       = ["Mike Campbell"]
+  spec.email         = ["mike@ydd.io"]
   spec.description   = %q{Enable Multi-tenant supported jobs to work with Sidekiq background worker}
-  spec.summary       = %q{Sidekiq support for Apartment}
-  spec.homepage      = "https://github.com/influitive/apartment-sidekiq"
+  spec.summary       = %q{Sidekiq support for Innkeeper}
+  spec.homepage      = "https://github.com/cpoms/innkeeper-sidekiq"
   spec.license       = "MIT"
 
   spec.files         = `git ls-files`.split($/)
@@ -22,6 +22,6 @@ Gem::Specification.new do |spec|
   spec.add_development_dependency "rake"
   spec.add_development_dependency 'minitest'
 
-  spec.add_dependency 'apartment', '>= 1.0'
+  spec.add_dependency 'innkeeper', '>= 1.0'
   spec.add_dependency 'sidekiq', '>= 2.11'
 end
