--- lib/apartment-sidekiq.rb
@@ -1,1 +1,0 @@
-require 'apartment/sidekiq'
