--- lib/apartment/sidekiq/middleware/honeybadger/server.rb
@@ -1,9 +1,0 @@
-module Apartment::Sidekiq::Middleware::Honeybadger
-  class Server
-    def call(worker_class, item, queue)
-      Honeybadger.context({tenant: Apartment::Tenant.current})
-
-      yield
-    end
-  end
-end
