--- lib/apartment/sidekiq/railtie.rb
@@ -1,7 +1,0 @@
-module Apartment::Sidekiq
-  class Railtie < Rails::Railtie
-    initializer "apartment.sidekiq" do
-      Apartment::Sidekiq::Middleware.run
-    end
-  end
-end
