--- lib/apartment/sidekiq/version.rb
@@ -1,5 +1,0 @@
-module Apartment
-  module Sidekiq
-    VERSION = "1.2.0"
-  end
-end
