+++ lib/innkeeper-sidekiq.rb
@@ -1,0 +1,1 @@
+require 'innkeeper/sidekiq'
