+++ lib/innkeeper/sidekiq/middleware/client.rb
@@ -1,7 +1,7 @@
-module Apartment::Sidekiq::Middleware
+module Innkeeper::Sidekiq::Middleware
   class Client
     def call(worker_class, item, queue, redis_pool=nil)
-      item["apartment"] ||= Apartment::Tenant.current
+      item["innkeeper"] ||= Innkeeper::Tenant.current
       yield
     end
   end
