+++ lib/innkeeper/sidekiq/middleware/honeybadger/server.rb
@@ -1,0 +1,9 @@
+module Innkeeper::Sidekiq::Middleware::Honeybadger
+  class Server
+    def call(worker_class, item, queue)
+      Honeybadger.context({tenant: Innkeeper::Tenant.current})
+
+      yield
+    end
+  end
+end
