+++ lib/innkeeper/sidekiq/railtie.rb
@@ -1,0 +1,7 @@
+module Innkeeper::Sidekiq
+  class Railtie < Rails::Railtie
+    initializer "innkeeper.sidekiq" do
+      Innkeeper::Sidekiq::Middleware.run
+    end
+  end
+end
