+++ lib/innkeeper/sidekiq/version.rb
@@ -1,0 +1,5 @@
+module Innkeeper
+  module Sidekiq
+    VERSION = "1.0.0"
+  end
+end
